+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * Refactored loader factory to return smart pointers to loaders ([#1597](https://github.com/CARTAvis/carta-backend/pull/1597)).
 
+### Changed
+* Refactored mapping of Stokes types to indices ([#1598](https://github.com/CARTAvis/carta-backend/pull/1598)).
+
 ## [6.0.0-beta.1]
 
 ### Fixed